Move inventory page components into inventory-page
