+++ src/app/inventory-page/D1Reputation.tsx
@@ -1,10 +1,10 @@
+import { D1Store } from 'app/inventory/store-types';
 import _ from 'lodash';
 import React from 'react';
 import { bungieNetPath } from '../dim-ui/BungieImage';
 import DiamondProgress from '../dim-ui/DiamondProgress';
 import PressTip from '../dim-ui/PressTip';
 import './D1Reputation.scss';
-import { D1Store } from './store-types';
 
 export default function D1Reputation({ store }: { store: D1Store }) {
   if (!store.progressions.length) {
