+++ src/app/inventory-page/D1ReputationSection.tsx
@@ -1,9 +1,9 @@
 import { t } from 'app/i18next-t';
+import { D1Store, DimStore } from 'app/inventory/store-types';
 import clsx from 'clsx';
 import React from 'react';
 import CollapsibleTitle from '../dim-ui/CollapsibleTitle';
 import D1Reputation from './D1Reputation';
-import { D1Store, DimStore } from './store-types';
 
 export default function D1ReputationSection({ stores }: { stores: DimStore[] }) {
   return (
