+++ src/app/inventory-page/DesktopStores.m.scss.d.ts
