+++ src/app/inventory-page/HeaderShadowDiv.m.scss
