+++ src/app/inventory-page/HeaderShadowDiv.m.scss.d.ts
