+++ src/app/inventory-page/HeaderShadowDiv.tsx
