+++ src/app/inventory-page/InventoryCollapsibleTitle.scss
