+++ src/app/inventory-page/InventoryCollapsibleTitle.tsx
@@ -1,5 +1,6 @@
 import { collapsedSelector } from 'app/dim-api/selectors';
 import { t } from 'app/i18next-t';
+import { DimStore } from 'app/inventory/store-types';
 import {
   postmasterAlmostFull,
   postmasterSpaceUsed,
@@ -14,7 +15,6 @@ import '../dim-ui/CollapsibleTitle.scss';
 import { toggleCollapsedSection } from '../settings/actions';
 import { AppIcon, collapseIcon, expandIcon } from '../shell/icons';
 import './InventoryCollapsibleTitle.scss';
-import { DimStore } from './store-types';
 
 interface Props {
   sectionId: string;
