+++ src/app/inventory-page/PhoneStores.tsx
@@ -1,18 +1,18 @@
 import { itemPop, scrollToPosition } from 'app/dim-ui/scroll';
 import { t } from 'app/i18next-t';
-import HeaderShadowDiv from 'app/inventory/HeaderShadowDiv';
+import { locateItem$ } from 'app/inventory/locate-item';
+import { DimStore } from 'app/inventory/store-types';
 import StoreStats from 'app/store-stats/StoreStats';
 import { useEventBusListener } from 'app/utils/hooks';
 import { wrap } from 'app/utils/util';
 import clsx from 'clsx';
 import { motion, PanInfo } from 'framer-motion';
 import React, { useCallback, useRef, useState } from 'react';
-import { InventoryBucket, InventoryBuckets } from './inventory-buckets';
-import { locateItem$ } from './locate-item';
+import { InventoryBucket, InventoryBuckets } from '../inventory/inventory-buckets';
+import { getCurrentStore, getStore, getVault } from '../inventory/stores-helpers';
+import HeaderShadowDiv from './HeaderShadowDiv';
 import PhoneStoresHeader from './PhoneStoresHeader';
-import { DimStore } from './store-types';
 import { StoreBuckets } from './StoreBuckets';
-import { getCurrentStore, getStore, getVault } from './stores-helpers';
 import './Stores.scss';
 
 interface Props {
