+++ src/app/inventory-page/PhoneStoresHeader.m.scss
