+++ src/app/inventory-page/PhoneStoresHeader.m.scss.d.ts
