+++ src/app/inventory-page/StoreBucket.scss
