+++ src/app/inventory-page/StoreBucketDropTarget.m.scss.d.ts
