+++ src/app/inventory-page/Stores.scss
